Move authentication method from repositories to services
