+++ app/repositories/user.js
@@ -1,7 +1,5 @@
 var mongoose = require( 'mongoose' );
-var jwt = require( 'jsonwebtoken' );
 var User = require( '../models/user.js' );
-var config = require( '../../config/config.js' );
 
 /**
  * This method is used to register new user with the given params
@@ -22,28 +20,4 @@ module.exports.createUser = function( userParams ) {
  */
 module.exports.findUserByEmail = function( email ) {
   return User.findOne( {email: email} );
-};
-
-/**
- * This method is used to authenticate user and return json web token
- *
- * @param {User} user - The user object which is found by the email
- * @param {String} userPassword - Password which must be checked for the authentication
- * @returns {Object}
- */
-module.exports.authenticateUser = function( user, userPassword ) {
-  if ( !user )
-    return {
-      success: false, message: 'User not found'
-    };
-
-  if ( user.password !== userPassword )
-    return {
-      success: false, message: 'Password does not match'
-    };
-
-  var token = jwt.sign( user, config.secret, {} );
-  return {
-    success: true, message: 'Token is ready', token: token
-  };
 };
